requirements.txt
# Core libraries
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.23.0

# API and data handling
requests>=2.28.0
python-dotenv>=1.0.0
pytz>=2023.3

# MTA-specific
protobuf>=3.20.0, <4.0.0
gtfs-realtime-bindings>=1.0.0
protobuf-to-dict>=0.1.0

# Visualization
plotly>=5.14.0
matplotlib>=3.7.0